Add GitHub Actions CI and tidy test code
Run fmt, clippy (-D warnings), and tests on push to main and on PRs,
with rust-cache and locked dependency resolution.

Drop unused encoding fields from golden fixture structs; serde ignores
the JSON keys by default. Fix clippy unnecessary_cast in roundtrip test.

+++ tests/roundtrip.rs
@@ -787,11 +787,7 @@ fn test_histogram_large_chunk() {
 
     for i in 0..50 {
         let t = base_time + i * 15000;
-        let h = create_test_histogram(
-            10 + i as u64,
-            18.4 + i as f64,
-            vec![4 + i as i64, 4 + i as i64],
-        );
+        let h = create_test_histogram(10 + i as u64, 18.4 + i as f64, vec![4 + i, 4 + i]);
         chunk.append(t, &h);
     }
